Move the Section filter to the Student column header
