+++ tool/src/java/org/sakaiproject/gradebookng/tool/pages/GradebookPage.html
@@ -34,12 +34,6 @@
           <li>
             <button id="toggleCategoriesToolbarItem">Group By Category</button>
           </li>
-          <li>
-            <select id="filterByGroup" wicket:id="groupFilter">
-              <option value="1">Section 1</option>
-              <option value="2">Section 2</option>
-            </select>
-          </li>
         </ul>
       </div>
       <table wicket:id="table" id="gradebookGradesTable" class="table-striped table-bordered table-hover"></table>
