+++ tool/src/java/org/sakaiproject/gradebookng/tool/panels/StudentNameColumnHeaderPanel.html
@@ -10,6 +10,10 @@
 		<form wicket:id="filterForm">
 			<input type="text" wicket:id="filter" wicket:message="placeholder:placeholder.students" />
 		</form>
+    <select id="filterByGroup" wicket:id="groupFilter">
+      <option value="1">Section 1</option>
+      <option value="2">Section 2</option>
+    </select>
 	</div>
 	
 	<!--  this comes from the wicket bootstrap module, needs to be sorted out. -->
