+++ tool/src/java/org/sakaiproject/gradebookng/tool/panels/StudentNameColumnHeaderPanel.java
@@ -16,6 +16,9 @@ import org.apache.wicket.model.IModel;
 import org.apache.wicket.model.Model;
 import org.apache.wicket.model.PropertyModel;
 import org.apache.wicket.model.ResourceModel;
+import org.apache.wicket.spring.injection.annot.SpringBean;
+import org.sakaiproject.gradebookng.business.GradebookNgBusinessService;
+import org.sakaiproject.gradebookng.business.model.GbGroup;
 import org.sakaiproject.gradebookng.business.model.GbStudentSortType;
 import org.sakaiproject.gradebookng.tool.model.StringModel;
 
@@ -30,6 +33,9 @@ public class StudentNameColumnHeaderPanel extends Panel {
 
 	private static final long serialVersionUID = 1L;
 
+  @SpringBean(name="org.sakaiproject.gradebookng.business.GradebookNgBusinessService")
+  protected GradebookNgBusinessService businessService;
+
 	public StudentNameColumnHeaderPanel(String id, GbStudentSortType currentSortOrder) {
 		super(id);
 		
@@ -42,9 +48,34 @@ public class StudentNameColumnHeaderPanel extends Panel {
 		//TODO use the list to render the dropdown, changing the text as appropriate
 		
 		add(new FilterForm("filterForm"));
-		
-		
-		
+
+
+    //section and group dropdown
+    final List<GbGroup> groups = this.businessService.getSiteSectionsAndGroups();
+
+    DropDownChoice<GbGroup> groupFilter = new DropDownChoice<GbGroup>("groupFilter", groups, new ChoiceRenderer<GbGroup>() {
+
+      private static final long serialVersionUID = 1L;
+
+      @Override
+      public Object getDisplayValue(GbGroup g) {
+        return g.getTitle();
+      }
+
+      @Override
+      public String getIdValue(GbGroup g, int index) {
+        return g.getId();
+      }
+
+    });
+
+    //TODO need to subclass the DDC to add the selectionchanged listener
+
+    groupFilter.setVisible(!groups.isEmpty());
+    groupFilter.setModel(new Model<GbGroup>()); //TODO update this so its aware of the currently selected filter. Maybe the form needs to maintain state and have this as a param?
+    groupFilter.setDefaultModelObject(groups.get(0)); //TODO update this
+    groupFilter.setNullValid(false);
+    add(groupFilter);
 	}
 	
 	private class FilterForm extends Form<StringModel> {
